refactor: moved everything to src folder
