+++ src/components/base/app-layout-grid-pagination-skeleton.vue
