+++ src/components/products/product-card-skeleton.vue
