+++ src/components/products/product-grid.vue
