+++ src/components/user/user-points-skeleton.vue
