+++ src/components/user/user-points.vue
