+++ tailwind.config.js
@@ -1,14 +1,14 @@
 /*
-** TailwindCSS Configuration File
-**
-** Docs: https://tailwindcss.com/docs/configuration
-** Default: https://github.com/tailwindcss/tailwindcss/blob/master/stubs/defaultConfig.stub.js
-*/
+ ** TailwindCSS Configuration File
+ **
+ ** Docs: https://tailwindcss.com/docs/configuration
+ ** Default: https://github.com/tailwindcss/tailwindcss/blob/master/stubs/defaultConfig.stub.js
+ */
 module.exports = {
   theme: {
     extend: {
       height: theme => ({
-        "80": "412px"
+        '80': '412px'
       }),
       colors: {
         primary: {
@@ -18,9 +18,9 @@ module.exports = {
           font: '#a3a3a3'
         },
         active: '#25bbf1'
-      },
-    },
+      }
+    }
   },
   variants: {},
   plugins: []
-}
+};
