+++ tsconfig.json
@@ -20,7 +20,7 @@
         "./*"
       ],
       "@/*": [
-        "./*"
+        "./src/*"
       ]
     },
     "types": [
